requirements.txt
langchain
langchain-openai
langchain-community
langchain-experimental
pandas
python-dotenv
streamlit